+++ routes/web.php
@@ -1,0 +1,5 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+
+Route::post(config('conekta.webhook'), 'WebhookController@handleWebhook')->name('webhook');